requirements.txt
# Core LangChain dependencies
langgraph>=0.2.0
langgraph-cli>=0.0.20
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0

# External API dependencies
tavily-python>=0.5.0
firecrawl-py>=1.0.0

# FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
sse-starlette>=1.6.5

# Utility dependencies
python-dotenv>=1.0.0

# Testing dependencies
pytest>=8.0.0
